fix: Overwrite sidePanelTemplate into FetchableDataGrid
